+++ Unity/Assets/Script/Roga2d/Interval/Roga2dScaleInterval.cs
@@ -3,7 +3,7 @@ using UnityEngine;
 public class Roga2dScaleInterval : Roga2dValueInterval<Vector2> {
 	private Roga2dNode node;
 	
-	public Roga2dScaleInterval(Roga2dNode node, Vector2 start, Vector2 end, int duration, bool tween) 
+	public Roga2dScaleInterval(Roga2dNode node, Vector2 start, Vector2 end, float duration, bool tween) 
 	: base(start, end, duration, tween) 
 	{
 		this.node = node;
